+++ Migrations/CHANGELOG.md
@@ -5,3 +5,4 @@
 - `dotnet ef database update`
 - `dotnet ef migrations add "AdicaoLider"`
 - `dotnet ef database update`
+- `dotnet ef migrations add "AdicionaSenha"`